requirements.txt
python-dotenv
requests  
langchain-google-genai
langchain
google-generativeai

